dirty hack to enable dynamic side-by-side operation
This really aught to be extended to all configuration params, but the
one I really need to make delta useful is the ability to situationally
do side-by-side diffs. Git does not make it possible to pass through
command line options to underlying tools, which is quite short-sighted.

This dirty hack:
- look for the DELTA_SIDE_BY_SIDE environment variable.
- If this env var is present and set to zero, override the side-by-side
  feature to false
- If this env var is present and set to a non-zero integer, override
  the side-by-side feature to true